refactor(security): reorganize SecurityConfig and enhance API documentation support
- Moved `SecurityConfig` to the `security` package for better structure.
- Updated public endpoint configuration to include Swagger and OpenAPI documentation paths.
- Adjusted package-info files to reflect the new structure and responsibilities.
