+++ src/main/java/com/krisnaajiep/expensetrackerapi/config/package-info.java
@@ -1,10 +1,10 @@
 /**
  * This package contains configuration classes. Configurations in this package include:
  * <ul>
- *   <li>Security configurations (Spring Security)</li>
  *   <li>Authentication configuration properties</li>
  *   <li>Rate limit and throttling configuration properties</li>
  *   <li>Filter configurations</li>
+ *   <li>Swagger configuration for API documentation</li>
  *   <li>Other application-wide settings and beans</li>
  * </ul>
  *
